fix(keep-awake): make the toggle real on Windows, and stop calling it a Mac
Three settings claimed to prevent sleep -- agents running, remote access
bound, a schedule armed -- and on Windows `platform_backend()` returned a
no-op, so all three counted their holds and asserted nothing. The labels
said "Mac" on top of that. `docs/windows-port-exclusions.md` exists so that
features are hidden rather than faked, and keep-awake is not listed there,
so this was a gap rather than a decision.

Windows has an exact equivalent, so it ships one: a power request, not
`SetThreadExecutionState`. The older call's effect belongs to the calling
thread and dies with it, and holds here are acquired and released from
wherever an agent, the remote host or the scheduler happens to be running --
so the thread that acquires is routinely not the one that releases. A power
request is process-wide and handle-based, and it carries the reason string,
so `powercfg /requests` names OxiMux and says why, the way `pmset -g
assertions` already did.

The backend token widened to 64 bits: macOS hands back a 32-bit assertion
id and Windows a HANDLE, and narrowing the handle to fit the smaller of the
two would truncate the pointer and leak the request. A test calls the real
kernel32 and asserts the round-tripped handle still names the request --
the failure it guards is invisible from inside the app and permanent from
outside it, since a clear that lands nowhere leaves a machine that simply
stops sleeping.

Copy across all five user-visible strings now says "this computer".

+++ apps/desktop/src/scheduler.rs
@@ -274,12 +274,12 @@ mod tests {
     }
 
     impl SleepAssertionBackend for MockBackend {
-        fn create(&self, _name: &str) -> Option<u32> {
+        fn create(&self, _name: &str) -> Option<u64> {
             let mut created = self.created.lock().unwrap();
             *created += 1;
-            Some(*created)
+            Some(u64::from(*created))
         }
-        fn release(&self, _id: u32) {}
+        fn release(&self, _id: u64) {}
     }
 
     /// A firer over an empty in-memory database, plus the backend so a test
